# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ErTraining",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "login",
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "leonardo.garuti.studio@fitstic-edu.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "LOGS OUT CURRENT LOGGED IN USER SESSION",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/usersList": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tuttor - Ritorna la lista di tutti gli utenti",
                "description": "Get the list of all users",
                "operationId": "usersList",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/tutorsList": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor - ritorna la lista di tutti i tutor",
                "description": "Get the list of tutors",
                "operationId": "tutorsList",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/teachersList": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor, ritorna la lista di tutti gli insegnanti",
                "description": "Get the list of teachers",
                "operationId": "teachersList",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/studentsList": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Ritorna la lista di tutti gli studenti",
                "description": "Get the list of students",
                "operationId": "studentsList",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/getUser/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor - ritorna le informazioni di un utente",
                "description": "Get the user by id",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/addUser": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor - crea un nuovo utente",
                "description": "Add user",
                "operationId": "addUser",
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "email",
                                    "password",
                                    "role_id"
                                ],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "format": "user_name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "format": "role_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/editUser/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor - modifica le informazioni di un utente",
                "description": "Edit user",
                "operationId": "editUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "email",
                                    "password",
                                    "role_id"
                                ],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "format": "user_name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/deleteUser/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Tutor - Cancella un'utente",
                "description": "delete the user by id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "user_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "email_verified_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "acces_token": {
                        "description": "token",
                        "type": "string",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "access_token"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}